% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_as_cols.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{sfc_as_cols
extract coordinates as new columns}
\usage{
sfc_as_cols(x, names = c("lon", "lat"))
}
\arguments{
\item{x}{sf dataframe}

\item{names}{Geometry column output names (by default lon and lat)}
}
\description{
sfc_as_cols
extract coordinates as new columns
}
\details{
source: source: https://github.com/r-spatial/sf/issues/231#issuecomment-282359896
}
